Moving other commands to celery tasks
